+++ CHANGES.md
@@ -17,6 +17,11 @@
   - @enm10k
 - [UPDATE] スポットライト接続時に spotlight_focus_rid / spotlight_unfocus_rid を指定できるようにする
   - @enm10k
+- [UPDATE] offer に mid が含まれる場合は、 mid を利用して sender を設定する
+  - @enm10k
+- [UPDATE] 依存ライブラリーのバージョンを上げる
+  - `com.android.tools.build:gradle` を 4.2.2 に上げる
+  - @enm10k
 - [ADD] データチャネルシグナリングに対応する
   - data_channel_signlaing, ignore_disconnect_websocket パラメータ設定を追加する
   - onDataChannel コールバックを実装する